s3server/auth/server/src/main/java/com/seagates3/util/ARNUtil.java
/*
 * COPYRIGHT 2016 SEAGATE LLC
 *
 * THIS DRAWING/DOCUMENT, ITS SPECIFICATIONS, AND THE DATA CONTAINED
 * HEREIN, ARE THE EXCLUSIVE PROPERTY OF SEAGATE TECHNOLOGY
 * LIMITED, ISSUED IN STRICT CONFIDENCE AND SHALL NOT, WITHOUT
 * THE PRIOR WRITTEN PERMISSION OF SEAGATE TECHNOLOGY LIMITED,
 * BE REPRODUCED, COPIED, OR DISCLOSED TO A THIRD PARTY, OR
 * USED FOR ANY PURPOSE WHATSOEVER, OR STORED IN A RETRIEVAL SYSTEM
 * EXCEPT AS ALLOWED BY THE TERMS OF SEAGATE LICENSES AND AGREEMENTS.
 *
 * YOU SHOULD HAVE RECEIVED A COPY OF SEAGATE'S LICENSE ALONG WITH
 * THIS RELEASE. IF NOT PLEASE CONTACT A SEAGATE REPRESENTATIVE
 * http://www.seagate.com/contact
 *
 * Original author:  Arjun Hariharan <arjun.hariharan@seagate.com>
 * Original creation date: 19-May-2016
 */
package com.seagates3.util;

public class ARNUtil {

    private static final String PARTITION = "seagate";
    private static final String SERVICE = "iam";

    public static String createARN(String accountId, String resource) {
        return String.format("arn:%s:%s::%s:%s", PARTITION, SERVICE,
                accountId, resource);
    }

    public static String createARN(String accountId, String resourceType,
            String resource) {
        return String.format("arn:%s:%s::%s:%s/%s", PARTITION, SERVICE,
                accountId, resourceType, resource);
    }
}
